+++ libm/mathf/ceilf.c
@@ -19,7 +19,7 @@ float ceilf(float x)
         if (j0 < 0) {  /* raise inexact if x != 0 */
             if (huge + x > (float)0.0) { /* return 0*sign(x) if |x|<1 */
                 if (i0 < 0) {
-                    i0 = 0x80000000;
+                    i0 = (int32_t)0x80000000;
                 } else if (!FLT_UWORD_IS_ZERO(ix)) {
                     i0 = 0x3f800000;
                 }
